Move the basic parser to its own module.
